requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0
scikit-learn>=1.0.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0
torch>=1.9.0
torchvision>=0.10.0
h5py>=3.3.0
tqdm>=4.62.0
notebook>=6.4.0
pandas>=1.3.0
PyYAML>=5.4.0
chromadb>=0.4.0 
fastapi
uvicorn[standard]
python-multipart 